Many problems of TotalView3d are fixed
The problem that a new slice in a 2D view showed old objects for totalview.isopaque==true is fixed.
The reason the 2D views showed differently rendered images for totalview.isopaque = true and false is identified.
The three 2D views are re-distributed for users to imagine a 3D box more easily.